Use nested tests for ClientUtils
